ansible: updated to 3.0.0
We're happy to announce that the ansible-3.0.0 package is now
available! This update is based on the ansible-base-2.10.x package
just like ansible-2.10 was so the changes shouldn't be too major.
However, it does contain new major versions of many collections which
means that there will be some backwards incompatible changes in the
modules and plugins.

If you would like to learn about how and why we got to version 3.0.0,
we invite you to visit our blog for the background as well as a Q&A:

- https://www.ansible.com/blog/announcing-the-community-ansible-3.0.0-package
- https://www.ansible.com/blog/ansible-3.0.0-qa