py-xdis: updated to 5.0.8
5.0.8:

Allow an alternate opmap - adds the capability to disassemble python bytecode that has
been frozen with a custom opcode mapping. This is particularly useful for disassembling
malware that uses custom opcode mappings in an attempt to hinder disassembly with standard
tools. The updates in this pull request are used by pydecipher, a tool to unfreeze and deob fuscate frozen python code.

Add Python versions 3.8.8 and 3.9.2