rust: normalize workarounds for cargo dependencies on NetBSD
Havard would like all rusts to be built with rust-cargo-static,
because this makes it easier to produce working bootstaps.

In order to do this, we need to handle the OpenSSL and curl
dependencies in older bootstrap kits properly.

This means, for the remaining bootstraps that do not yet have
cargo with static/vendored dependencies:

- depend on base 8.0 openssl (compat80) on i386.
- depend on pkgsrc curl on aarch64 and earmv7.
  the armv7 bootstrap is built for 9.0 so does not need compat80.
  the aarch64 bootstrap is built for 8.99.50 so does not need compat80.