gopls: update to 0.6.9.
0.6.9
This is a patch release to fix golang/go#45092.

0.6.8
This is a patch release to fix golang/go#45075.

0.6.7
New analyzers

The unusedwrite and nilness analyzers are now available, but they
are disabled by default. Note that enabling them permanently may
significantly increase memory usage. You can configure analyzers
through the analyses setting.  Improved hover with time.Duration

Hover for any constant of type time.Duration is shown both in
nanoseconds and as a human-readable string.

Fixes

A full list of all issues fixed can be found in the gopls/v0.6.7
milestone. To report a new problem, please file a new issue at
https://golang.org/issues/new.