sqlite3: updated to 3.35.4
3.35.4

Fix a defect in the query planner optimization identified by item 8b above.
Fix a defect in the new RETURNING syntax.
Fix the new RETURNING feature so that it raises an error if one of the terms in the RETURNING clause references a unknown table, instead of silently ignoring that error.
Fix an assertion associated with aggregate function processing that was incorrectly triggered by the push-down optimization.