chat/coyim: Applies the correct fix for gotk3 from upstream.
As of Golang 1.16, the cgo tool will no longer try to translate C struct
bitfields into Go struct fields, even if their size can be represented
in Go. The order in which C bitfields appear in memory is implementation
dependent, so in some cases the cgo tool produced results that were
silently incorrect.

In this case "accel_flags" is the bitwise field in question. A new
declaration for GtkAccelKey structure that does not explicitly access
to 'accel_flags' has been addressed in the patch.

https://github.com/gotk3/gotk3/pull/730

Patch will be removed when vendor library will be updated.